requirements.txt
python-telegram-bot==20.3
aiohttp
Flask==3.0.0
nest-asyncio==1.6.0
waitress==2.1.2
